requirements.txt
autopep8==1.6.0
azure-cosmos==4.3.0b4
fastapi==0.78.0
funmark==0.1.1
matplotlib==3.5.2
mpmath==1.2.1
numpy==1.21.6
pandas==1.3.5
python-dotenv==0.20.0
uvicorn==0.17.6
